backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
crispy-bootstrap5==0.7
Django==4.1.7
django-bootstrap5==22.2
django-crispy-forms==2.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
